vlib:  Move bytes repetition logic to vbytes.cc
