+++ v/vlib/vlib/types/packed.cc
@@ -199,6 +199,9 @@ namespace vlib
 				
 				return binary_bitwise_op( op, a, b.string().data() );
 			
+			case Op_multiply:
+				return multiply( (const VBytes&) a, b );
+			
 			case Op_divide:
 				return division( (const VBytes&) a, b );
