+++ v/vlib/vlib/types/vbytes.hh
@@ -34,6 +34,7 @@ namespace vlib
 	
 	Value assign_byte_to_index( Value& s, const Value& x, const Byte& byte );
 	
+	Value multiply( const VBytes& bytes, const Value& factor  );
 	Value division( const VBytes& bytes, const Value& divisor );
 	
 	Value vbytes_contains( const Value& a, const Value& b );
